requirements.txt
Flask==3.0.3
google-generativeai==0.8.3
google-api-core==2.21.0
google-auth==2.35.0
google-api-python-client==2.149.0
